# housingserver-specification/src/main/resources/openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080/api/v2/housing
    description: Local development server
info:
  description: Housing benefit application handling demo API
  version: "1.0.0"
  title: Housing benefit api API
  contact:
    email: riku.sarlin@gmail.com
  license:
    name: MIT
    url: 'https://mit-license.org/'
...
paths:
...
  /application/{caseId}:
    get:
      tags:
        - developers
      summary: fetch application of a given case
      operationId: fetchApplication
      description: |
        Fetch all incomes of a given case
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitApplication'
    put:
      tags:
        - developers
      summary: update application in a given case
      operationId: updateApplication
      description: Updates application info
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: updated application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitApplication'
        '400':
          description: update not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HousingBenefitApplication'
        description: Application to update
    delete:
      tags:
        - admins
      summary: delete application of a given case
      operationId: deleteApplication
      description: Deletes application of a given case
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '204':
          description: application deleted successfully
        '400':
          description: delete not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: application not found
    post:
      tags:
        - developers
      summary: add application to case
      operationId: addApplication
      description: Adds an application to a given case
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '201':
          description: application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitApplication'
        '400':
          description: application creation not ok, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HousingBenefitApplication'
        description: Application to add
  /application/{caseId}/check:
    get:
      tags:
        - developers
      summary: cross-check application info in a housing benefit application
      operationId: checkApplication
      description: |
        Cross validate by income id
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: application validation ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitApplication'
        '400':
          description: application validation not ok, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: application not found
...
  /case/{caseId}:
    get:
      tags:
        - developers
      summary: fetch housing benefit case
      operationId: fetchHousingBenefitCaseById
      description: |
        Fetch housing benefit case by id
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: housing benefit application found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitCase'
        '400':
          description: fetch not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: housing benefit case not found
    put:
      tags:
        - developers
      summary: update housing benefit case
      operationId: updateHousingBenefitCase
      description: Updates housing benefit case
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: updated housing benefit case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitCase'
        '400':
          description: update not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HousingBenefitCase'
        description: Housing benefit case to update
    delete:
      tags:
        - admins
      summary: delete housing benefit case
      operationId: deleteHousingBenefitCase
      description: Deletes housing benefit case
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '204':
          description: housing benefit case deleted successfully
        '400':
          description: delete not successful, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: housing benefit case not found
  /case/{caseId}/check:
    get:
      tags:
        - developers
      summary: cross-check housing benefit case
      operationId: checkHousingBenefitCase
      description: |
        Cross validate housing benefit case
      parameters:
        - in: path
          name: caseId
          description: id of case
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: housing benefit case validation ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitCase'
        '400':
          description: housing benefit case validation not ok, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: housing benefit case not found
  /case:
    post:
      tags:
        - developers
      summary: add housing benefit case based on application
      operationId: addHousingBenefitCase
      description: Adds housing benefit case
      responses:
        '201':
          description: housing benefit case nad application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingBenefitCase'
        '400':
          description: housing benefit case and application creation not ok, error message returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHousingBenefitCase'
        description: housing benefit case and application to add
components:
  schemas:
    ErrorMessage:
      type: object
      description: Info about error(s) in backend processing
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: number
          format: integer
          description: HTTP status code
        error:
          type: string
          description: description of error
        path:
          type: string
          description: URL path of error
        message:
          type: array
          items:
            type: object
            properties:
              fieldName:
                type: string
                description: name of field, if applicable, where error occured
              errorMessage:
                type: string
                description: a more detailed description of the error
              errorCode:
                type: string
                description: error code
    Id:
      type: integer
      description: Unique identifier of a resource
    IdAndDateRange:
      type: object
      description: Id and possibly open range between two dates - that is, both start end end dates can be empty
      properties:
        id:
          $ref: '#/components/schemas/Id'
        startDate:
          type: string
          format: date
          example: '2020-10-08'
        endDate:
          type: string
          format: date
          example: '2020-10-08'
    Person:
      type: object
      description: A natural person residing in Finland
      properties: 
        id:
          $ref: '#/components/schemas/Id'
        personNumber:
          $ref: '#/components/schemas/PersonNumber'
        firstName:
          type: string
          description: First name of a person
          maxLength: 80
        lastName:
          type: string
          description: Last name of a person
          maxLength: 80
        birthDate:
          type: string
          format: date
          example: '1971-08-71'
    HouseholdMember:
      type: object
      description: Household member info
      allOf:
          - $ref: '#/components/schemas/IdAndDateRange'
          - type: object
            properties: 
              person:
                $ref: '#/components/schemas/Person'
    PersonNumber:
      type: string
      description: Finnish person number
      minLength: 11
      maxLength: 11
      pattern: '^\d{6}[A+-]\d{3}[0-9A-FHJ-NPR-Y]$'
    ExpenseType:
      type: string
      enum: 
        - RENT
        - ELECTRICITY
        - WATER
        - INTEREST
        - MAINTENANCE_CHARGE
        - OTHER
    IncomeType:
      type: string
      enum:
        - SALARY
        - PENSION
        - SOCIAL_BENEFIT
        - STUDY_GRANT
        - DIVIDEND
        - INTEREST
        - OTHER
    Income:
      type: object
      description: Income info
      allOf:
          - $ref: '#/components/schemas/IdAndDateRange'
          - type: object
            properties: 
              incomeType:
                $ref: '#/components/schemas/IncomeType'
              amount:
                type: number
                format: double
              otherIncomeDescription:
                type: string
                description: A textual description for income if income type is OTHER
    Expense:
      type: object
      description: Expense info
      allOf:
          - $ref: '#/components/schemas/IdAndDateRange'
          - type: object
            properties: 
              expenseType:
                $ref: '#/components/schemas/ExpenseType'
              amount:
                type: number
                format: double
              otherExpenseDescription:
                type: string
                description: A textual description for expense if expense type is OTHER
    HousingBenefitApplication:
      type: object
      description: Housing benefit application info
      allOf:
          - $ref: '#/components/schemas/IdAndDateRange'
          - type: object
            properties:
              applicant:
                $ref: '#/components/schemas/Person'
              received:
                type: string
                format: date-time
    CaseState:
      type: string
      enum:
        - NEW
        - ACTIVE
        - WAITING
        - READY
        - CANCELED
    HousingBenefitCase:
      type: object
      description: Housing benefit case
      properties:
        id:
          $ref: '#/components/schemas/Id'
        caseState:
          $ref: '#/components/schemas/CaseState'
        customer:
          $ref: '#/components/schemas/Person'
        housingBenefitApplication:
          $ref: '#/components/schemas/HousingBenefitApplication'
        householdMembers:
          type: array
          items:
            $ref: '#/components/schemas/HouseholdMember'
        housingExpenses:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        incomes:
          type: array
          items:
            $ref: '#/components/schemas/Income'
    NewHousingBenefitCase:
      type: object
      description: Housing benefit case
      properties:
        id:
          $ref: '#/components/schemas/Id'
        caseState:
          $ref: '#/components/schemas/CaseState'
        customer:
          $ref: '#/components/schemas/Person'
        housingBenefitApplication:
          $ref: '#/components/schemas/HousingBenefitApplication'